==> notebooks/1. DPP 05 - GPI 14.ipynb <==
# Header

import os

## Notebook Configuration

# path to the this notebook
# NOTE: Replace this with your project path if needed
PROJECT_PATH = (
    "/content/drive/My Drive/Colab Notebooks"
    if "google.colab" in str(get_ipython())
    else "."
)

# path to the data folder
# NOTE: Replace this with your data path if needed
DATA_PATH = f"{PROJECT_PATH}/data" if "google.colab" in str(get_ipython()) else PROJECT_PATH
# NOTE: For colab we use content so it doesn't load on google drive storage
RAW_DATA_PATH = f"{PROJECT_PATH}/data" if "google.colab" in str(get_ipython()) else f"{PROJECT_PATH}/data"

## Colab Setup

if "google.colab" in str(get_ipython()):
    from google.colab import drive
    drive.mount('/content/drive')

    # setup libraries used by notebook
    #os.system("pip install -q kaggle")

os.chdir(PROJECT_PATH)

## Library Import

import itertools
import json
import requests
import shutil
import typing
import zipfile

from io import BytesIO
from pathlib import Path

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd

from tqdm import tqdm

---

# Data Load

ndc_df = pd.read_parquet(f"{DATA_PATH}/preprocessed/comp_ndc.parquet")
packaging_df = pd.read_parquet(f"{DATA_PATH}/preprocessed/comp_packaging.parquet")
generic_name_df = pd.read_parquet(f"{DATA_PATH}/preprocessed/comp_generic_name.parquet")
active_ingredients_df = pd.read_parquet(f"{DATA_PATH}/preprocessed/comp_active_ingredients.parquet")

# Data Pre-Processing

ndc_directory = (
    packaging_df[["product_ndc", "package_ndc", "ndc"]].drop_duplicates()
    .merge(
        ndc_df[["product_ndc", "dosage_form"]]
        .assign(dosage_form=lambda f: f["dosage_form"].str.strip().str.upper())
        .drop_duplicates(),
        on=["product_ndc"],
        how='left'
    )
    .merge(
        generic_name_df[["product_ndc", "generic_name"]]
        .assign(generic_name=lambda f: f["generic_name"].str.strip().str.upper())
        .drop_duplicates(),
        on=["product_ndc"],
        how='left'
    )
    .merge(
        active_ingredients_df[["product_ndc", "name", "strength"]]
        .rename(columns={"name": "generic_name"})
        .assign(generic_name=lambda f: f["generic_name"].str.strip().str.upper())
        .drop_duplicates(),
        on=["product_ndc", "generic_name"],
        how='left'
    )
)

# adjust strength
nulls_to_add = ndc_directory.loc[lambda f: f["strength"].isnull()].drop(columns=["strength"]).merge(
  active_ingredients_df[["product_ndc", "strength"]].drop_duplicates(),
  on=["product_ndc"],
  how='left'
)
ndc_directory = pd.concat([ndc_directory.loc[lambda f: f["strength"].notnull()], nulls_to_add])

# Create a unique identifier for each row
ndc_directory["gpi14"] = (
    ndc_directory['generic_name'].fillna("N/A")
    + " | "
    + ndc_directory['dosage_form'].fillna("N/A")
    + " | "
    + ndc_directory['strength'].fillna("N/A")
)

# Use factorize to assign unique integer IDs to the unique combinations
ndc_directory['gpi14_id'] = pd.factorize(ndc_directory['gpi14'])[0]

#ndc_directory = ndc_directory.drop(columns=["dosage_form", "generic_name", "strength"])

ndc_directory.loc[lambda f: f["product_ndc"] == "41701-013"]

packaging_df.sample(10).values

packaging_df.loc[lambda f: f["product_ndc"] == "41701-013"].values[0]

# Export

ndc_directory.to_parquet(f"{DATA_PATH}/preprocessed/gpi14.parquet", index=False)

---
